Log 'Begin poll(*)' before every poll.
Users should typically write a debug message before polling. Do this
for the user.

Message does not log all the arguments to poll(). That would be a very
verbose message. I think logging enough to workout how long the poll()
could take is probably good enough.

In the current implementation, there isn't a way for a user to not log
this message. This is probably fine because logging provides quite a
lot of configuration flexibility. However, this might need to be
reconsider if some users complain.